move diagnostic test files into own folder
