+++ packages/language-server/test/plugins/typescript/testfiles/diagnostics/diagnostics-directive-types.svelte
