+++ packages/language-server/test/plugins/typescript/testfiles/diagnostics/diagnostics-js-notypecheck.svelte
