+++ packages/language-server/test/plugins/typescript/testfiles/diagnostics/diagnostics-slots.svelte
