+++ packages/language-server/test/plugins/typescript/testfiles/diagnostics/diagnostics.svelte
