+++ packages/language-server/test/plugins/typescript/testfiles/diagnostics/tsconfig.json
@@ -1,0 +1,10 @@
+{
+    "compilerOptions": {
+        "strict": true,
+        /**
+          This is actually not needed, but makes the tests faster
+          because TS does not look up other types.
+        */
+        "types": ["svelte"]
+    }
+}
